Refactor test fixtures for direct AudioTag and Plex API use
Because:
- Indirect track_factory and track fixtures reduced test clarity
- Plex API/player mocks were duplicated and inconsistent

Removes the 'track' fixture and all track_factory dependencies from
test_id3_handler.py, using explicit AudioTag construction. Adds shared,
comprehensive Plex API/player mocking fixtures to conftest.py, and updates
test_plex_player.py to use these, removing local duplicates and renaming
parameters for clarity. No breaking changes; test coverage and behavior
unchanged. All affected tests now use direct, explicit values for track
objects and standardized Plex mocks, improving readability, fixture hygiene,
and maintainability across the test suite.

Affects:
- tests/filesystem_provider/test_id3_handler.py
- tests/conftest.py
- tests/mediaplayer/test_plex_player.py
